backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# API Documentation
drf-spectacular>=0.27.0

# Environment Variables
python-decouple>=3.8

# CORS Headers (for frontend integration)
django-cors-headers>=4.3.0

# Production Server
gunicorn>=21.2.0

# Static Files (Production)
whitenoise>=6.6.0

# Utilities
pytz>=2023.3

